LPS-58623 ee-6.2.x: We should keep other users membership(users come from ladp server2 or create in portal) of usergroup when synchronize ldap1 server

+++ portal-impl/src/com/liferay/portal/security/ldap/PortalLDAPImporterImpl.java
@@ -1077,7 +1077,9 @@ public class PortalLDAPImporterImpl implements PortalLDAPImporter {
 			userGroupId);
 
 		for (User user : userGroupUsers) {
-			if (!newUserIds.contains(user.getUserId())) {
+			if ((ldapServerId == user.getLdapServerId()) &&
+				!newUserIds.contains(user.getUserId())) {
+
 				UserLocalServiceUtil.deleteUserGroupUser(
 					userGroupId, user.getUserId());
 			}
